test: update ITDatastoreTest test to attempt multiple times if a test fails
Datastore (when not Firestore in Datastore mode) is eventually
consistent. Sometimes this means queries may not show the expected
results within the short timespan of a test. This change adds a new
JUnit Rule - MultipleAttemptsRule - which allows tests to attempt to
succeed up to three times. If a test is unable to pass within the 3
allowed attempts, the errors from each attempt will be returned as an
overall failure.

Fixes #97
Fixes #101